ARM: tegra: implement DT copying/deleting when booting kernel
In some configurations, U-Boot is chain-loaded by another bootloader, and
that bootloader passes a DTB to the next stage. Generally, that DTB is
intended for consumption by an NVIDIA L4T kernel, and hence uses non-
upstreamed DT bindings, and hence is not suitable for consumption by
U-Boot as a control DTB. However, the other bootloader has often placed
useful data into that DTB, such as SDRAM training data or L4T-specific
board enumeration data, and this data must be passed on to any L4T kernel
that is eventually booted.

This patch implements a feature that allows copying a set of DT nodes and
properties from that DTB into the DTB that's passed to the kernel. The
list of nodes/properties to copy is taken from environment variables,
which allows the user to configure or disable this operation. Future
patches will configure the default environment to contain the required
copying configuration on a per-SoC basis.

Equally, there are times when certain nodes/properties must be deleted
from the DTB that's passed to the kernel. This patch also implementes
that, again configured via environment variables.

Bug 200211165

Change-Id: Ibea20f204b42dc9ecad6d02a2600b82b88bb0b8a
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/1195409
Reviewed-by: Tom Warren <twarren@nvidia.com>
Signed-off-by: Thog <me@thog.eu>