ARM: tegra: fix TX1 camera board ID EEPROM failure
When the code to enable power to the camera board EEPROM was
ported from rel-24 in commit 93794875, the CONFIG switch used
was not changed from CONFIG_OF_ADD_CAM_BOARD_ID to
CONFIG_NV_BOARD_ID_EEPROM_CAM. The old CONFIG switch was still
being used in p2371-2180.c, which resulted in the
camera_board_power_init() code not being compiled in and the
camera board ID EEPROM could not be found/read. This resulted
in an error message in U-Boot, and the plugin-manager info
for the camera board wasn't populated in the DTB passed to the
kernel.

Fixed the CONFIG switch use so it's all consistent now, and
added an error printf in the board ID read function to show
the I2C bus/dev address on failure to help w/future debug.

Bug 200217895

Change-Id: Ib86e4d16ad457ab22bbdea2efb4825ffe7da277d
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: http://git-master/r/1232055
Reviewed-by: Stephen Warren <swarren@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shreshtha Sahu <ssahu@nvidia.com>
Tested-by: Shreshtha Sahu <ssahu@nvidia.com>
Signed-off-by: Thog <me@thog.eu>