ARM: tegra: Add m3402-0000 (OrcaRacer) board support
Orca Racer is a M3402 CPU board married to a E3401 baseboard. The
combination contains a T210 SoC, 4G LPDDR4, Max20024 PMIC, mSATA,
5x I2C, 4x SPI, QSPI, 2x USB2.0 and 1x USB3.0. A E2603/E2606 display
module (eDP/DP) can be connected to the baseboard.

This is a basic port that boots to the U-Boot command line. It will
need more work (SATA/AHCI driver, PMIC support, etc.) to be able to
boot a kernel/dtb from mSATA.

Bug 1842054

Change-Id: I0a1afc39c2bf8c90c793661c7d03b6083921132e
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: http://git-master/r/1263994
Signed-off-by: Thog <me@thog.eu>