ARM: tegra: skip DT copy if src==dst
Soon, the user will be able to pass the cboot-supplied DT to the kernel
rather than loading a new DT. In this scenario, the DT copy routines will
attempt to copy to and from the same DTB. Detect this case, and skip the
copy operation.

Bug 1831386.

Change-Id: I6401014b9dbcc64460714c4c48b1e78875245ff8
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/1241737
Reviewed-by: Tom Warren <twarren@nvidia.com>
GVS: Gerrit_Virtual_Submit
Signed-off-by: Thog <me@thog.eu>